services/ml-recommendation/requirements.txt
fastapi
uvicorn
pandas
scikit-learn
joblib
psycopg2-binary